The training and testing datasets respectively.
This is a logfile from an intrusion detection system (IDS). It contains records in the format: "time","sourceIP", "sourcePort", "destIP", "destPort", "classification", "priority", "label", "packet info", "packet info cont'd", "xref". We will be concerned with the sourceIP, destIP and classication fields for the purpose of this demonstration.